feat(lm): initalize components for local llm config in app settings
